requirements.txt
# === Core utilities ===
pandas
openpyxl
requests
beautifulsoup4
python-dotenv
tqdm
pymongo[srv]

# === CLI ===
click

# === Vector store (Qdrant) ===
qdrant-client

# === LLM embeddings ===
sentence-transformers
torch  # required for sentence-transformers

# === Gemini API ===
google-generativeai
langchain-google-genai

# === Orchestration + Evaluation (Optional for production) ===
langgraph
ragas
arize-phoenix
llama-index

# === Reranker and Transformers support ===
transformers
accelerate

# === API server ===
fastapi
uvicorn

# === OpenAI function calling and LangGraph agent support ===
openai

# === Auth ===
pyjwt
bcrypt

